Consolidate x509's into a single file. No need to have a file per (mutually exclusive) function.
